fbcon: fix integer overflow in font allocation
systemd-vconsole-setup service fails with error code 71 due
to integer overflow in font allocation. This issue is only
reproducible on OL7 systems. Fix it by using size_t instead of
signed int.

Orabug: 38702507

+++ drivers/video/fbdev/core/fbcon.c
@@ -2482,7 +2482,7 @@ static int fbcon_set_font(struct vc_data *vc, struct console_font *font,
 	unsigned charcount = font->charcount;
 	int w = font->width;
 	int h = font->height;
-	int size, alloc_size;
+	size_t size, alloc_size;
 	int i, csum;
 	u8 *new_data, *data = font->data;
 	int pitch = PITCH(font->width);
